divided note and noteList routers
